Book files refactored
