+++ book-club-api/src/author/dto/update-author.dto.ts
@@ -1,4 +1,4 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateAuthorDto } from './create-author.dto';
+// import { PartialType } from '@nestjs/mapped-types';
+// import { CreateAuthorDto } from './create-author.dto';
 
-export class UpdateAuthorDto extends PartialType(CreateAuthorDto) {}
+// export class UpdateAuthorDto extends PartialType(CreateAuthorDto) {}
