+++ book-club-api/src/gender/dto/create-gender.dto.ts
@@ -1,1 +1,1 @@
-export class CreateGenderDto {}
+// export class CreateGenderDto {}
