+++ book-club-api/src/gender/dto/update-gender.dto.ts
@@ -1,4 +1,4 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateGenderDto } from './create-gender.dto';
+// import { PartialType } from '@nestjs/mapped-types';
+// import { CreateGenderDto } from './create-gender.dto';
 
-export class UpdateGenderDto extends PartialType(CreateGenderDto) {}
+// export class UpdateGenderDto extends PartialType(CreateGenderDto) {}
